+++ src/Components/Button.php
@@ -5,10 +5,12 @@ namespace Sikessem\UI\Components;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Sikessem\UI\Common\FormControl;
-use Sikessem\UI\UIFacade;
+use Sikessem\UI\Concerns\HandlesComponentAlias;
 
 class Button extends FormControl
 {
+    use HandlesComponentAlias;
+
     public ?string $href = null;
 
     /**
@@ -41,11 +43,9 @@ class Button extends FormControl
 
     public function render(): View|Factory
     {
-        $prefix = preg_quote(UIFacade::prefix(), '/');
-        $suffix = 'button';
-        if (preg_match("/$prefix-(?P<alias>.*?)-$suffix/", $this->componentName, $matches)) {
-            $this->type = $matches['alias'];
-        }
+        $this->handleAlias(function (string $alias) {
+            $this->type = $alias;
+        });
 
         return view('ui::components.button');
     }